Backend/Model/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pdfplumber==0.9.0
python-docx==0.8.11
spacy==3.7.1
lxml==4.9.3
